Fixes and refactor of TrackingUtils
